[HUDI-441] Rename WorkflowDagGenerator and some class names in test package
