+++ hudi-test-suite/src/main/java/org/apache/hudi/testsuite/dag/WorkflowDagGenerator.java
@@ -18,58 +18,14 @@
 
 package org.apache.hudi.testsuite.dag;
 
-import org.apache.hudi.common.util.collection.Pair;
-import org.apache.hudi.testsuite.configuration.DeltaConfig.Config;
-import org.apache.hudi.testsuite.dag.nodes.DagNode;
-import org.apache.hudi.testsuite.dag.nodes.HiveQueryNode;
-import org.apache.hudi.testsuite.dag.nodes.InsertNode;
-import org.apache.hudi.testsuite.dag.nodes.UpsertNode;
-
-import java.util.ArrayList;
-import java.util.List;
-
 /**
- * An example of how to generate a workflow dag programmatically. This is also used as the default workflow dag if
- * none is provided.
+ * A interface which represents a workflow dag generator.
  */
-public class WorkflowDagGenerator {
+public interface WorkflowDagGenerator {
 
-  public WorkflowDag build() {
-
-    DagNode root = new InsertNode(Config.newBuilder()
-        .withNumRecordsToInsert(100)
-        .withNumInsertPartitions(1)
-        .withNumTimesToRepeat(2)
-        .withRecordSize(1000).build());
-
-    DagNode child1 = new InsertNode(Config.newBuilder()
-        .withNumRecordsToInsert(100)
-        .withNumInsertPartitions(1)
-        .withNumTimesToRepeat(2)
-        .withRecordSize(1000).build());
-
-    root.addChildNode(child1);
-
-    DagNode child1OfChild1 = new UpsertNode(Config.newBuilder()
-        .withNumRecordsToUpdate(100)
-        .withNumUpsertPartitions(2)
-        .withNumTimesToRepeat(1)
-        .withRecordSize(1000).build());
-
-    // Tests running 2 nodes in parallel
-    child1.addChildNode(child1OfChild1);
-
-    List<Pair<String, Integer>> queryAndResult = new ArrayList<>();
-    queryAndResult.add(Pair.of("select " + "count(*) from testdb1.hive_trips group "
-        + "by rider having count(*) < 1", 0));
-    DagNode child2OfChild1 = new HiveQueryNode(Config.newBuilder()
-        .withHiveQueryAndResults(queryAndResult).withHiveLocal(true).build());
-    child1.addChildNode(child2OfChild1);
-
-    List<DagNode> rootNodes = new ArrayList<>();
-    rootNodes.add(root);
-
-    return new WorkflowDag(rootNodes);
-  }
+  /**
+   * Builds a {@link WorkflowDag}.
+   */
+  WorkflowDag build();
 
 }
