+++ hudi-test-suite/src/test/java/org/apache/hudi/testsuite/job/TestHoodieTestSuiteJob.java
@@ -23,9 +23,9 @@ import org.apache.hudi.common.table.HoodieTableMetaClient;
 import org.apache.hudi.common.util.TypedProperties;
 import org.apache.hudi.testsuite.DeltaInputFormat;
 import org.apache.hudi.testsuite.DeltaOutputType;
-import org.apache.hudi.testsuite.dag.TestHiveSyncDag;
-import org.apache.hudi.testsuite.dag.TestInsertOnlyDag;
-import org.apache.hudi.testsuite.dag.TestInsertUpsertDag;
+import org.apache.hudi.testsuite.dag.HiveSyncDagGenerator;
+import org.apache.hudi.testsuite.dag.InsertOnlyDagGenerator;
+import org.apache.hudi.testsuite.dag.InsertUpsertDagGenerator;
 import org.apache.hudi.testsuite.dag.WorkflowDagGenerator;
 import org.apache.hudi.testsuite.job.HoodieTestSuiteJob.HoodieTestSuiteConfig;
 import org.apache.hudi.utilities.UtilitiesTestBase;
@@ -134,7 +134,7 @@ public class TestHoodieTestSuiteJob extends UtilitiesTestBase {
     String inputBasePath = dfsBasePath + "/input";
     String outputBasePath = dfsBasePath + "/result";
     HoodieTestSuiteConfig cfg = makeConfig(inputBasePath, outputBasePath);
-    cfg.workloadDagGenerator = TestInsertOnlyDag.class.getName();
+    cfg.workloadDagGenerator = InsertOnlyDagGenerator.class.getName();
     HoodieTestSuiteJob hoodieTestSuiteJob = new HoodieTestSuiteJob(cfg, jsc);
     hoodieTestSuiteJob.runTestSuite();
     HoodieTableMetaClient metaClient = new HoodieTableMetaClient(new Configuration(), cfg.targetBasePath);
@@ -148,7 +148,7 @@ public class TestHoodieTestSuiteJob extends UtilitiesTestBase {
     String inputBasePath = dfsBasePath + "/input/" + UUID.randomUUID().toString();
     String outputBasePath = dfsBasePath + "/result/" + UUID.randomUUID().toString();
     HoodieTestSuiteConfig cfg = makeConfig(inputBasePath, outputBasePath);
-    cfg.workloadDagGenerator = TestInsertUpsertDag.class.getName();
+    cfg.workloadDagGenerator = InsertUpsertDagGenerator.class.getName();
     HoodieTestSuiteJob hoodieTestSuiteJob = new HoodieTestSuiteJob(cfg, jsc);
     hoodieTestSuiteJob.runTestSuite();
     HoodieTableMetaClient metaClient = new HoodieTableMetaClient(new Configuration(), cfg.targetBasePath);
@@ -176,7 +176,7 @@ public class TestHoodieTestSuiteJob extends UtilitiesTestBase {
     String inputBasePath = dfsBasePath + "/input";
     String outputBasePath = dfsBasePath + "/result";
     HoodieTestSuiteConfig cfg = makeConfig(inputBasePath, outputBasePath);
-    cfg.workloadDagGenerator = TestHiveSyncDag.class.getName();
+    cfg.workloadDagGenerator = HiveSyncDagGenerator.class.getName();
     HoodieTestSuiteJob hoodieTestSuiteJob = new HoodieTestSuiteJob(cfg, jsc);
     hoodieTestSuiteJob.runTestSuite();
     HoodieTableMetaClient metaClient = new HoodieTableMetaClient(new Configuration(), cfg.targetBasePath);
